# services/core/src/main/resources/META-INF/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "PageUserView" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserView"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
      "UpdateMe" : {
        "type" : "object",
        "required" : [ "displayName" ],
        "properties" : {
          "displayName" : {
            "type" : "string",
            "pattern" : "\\S",
            "maxLength" : 255
          }
        }
      },
...
      "UpdateUser" : {
        "type" : "object",
        "properties" : {
          "role" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string",
            "maxLength" : 255
          }
        }
      },
      "UserView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "email" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "tenantId" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Authentication"
      }
    }
  },
  "paths" : {
...
    "/api/platform/v1/users" : {
      "get" : {
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageUserView"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List",
        "tags" : [ "User Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "owner", "admin" ]
        } ]
      }
    },
    "/api/platform/v1/users/me" : {
      "patch" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateMe"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserView"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Update Me",
        "tags" : [ "User Resource" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserView"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Me",
        "tags" : [ "User Resource" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/api/platform/v1/users/{id}" : {
      "patch" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateUser"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserView"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Update",
        "tags" : [ "User Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "owner", "admin" ]
        } ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserView"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get",
        "tags" : [ "User Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "owner", "admin" ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Delete",
        "tags" : [ "User Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "owner", "admin" ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "appgrove core API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}